devhost: refactor management script into own python package
preperation for PL-133536
