+++ nixos/roles/devhost/fc.devhost/pyproject.toml
@@ -1,0 +1,27 @@
+[tool.black]
+line-length = 80
+
+[tool.isort]
+profile = "black"
+line_length = 80
+
+[tool.ruff]
+line-length = 80
+
+[build-system]
+requires = ["setuptools >= 61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "fc.devhost"
+dynamic = ["version"]
+requires-python = ">= 3.11"
+dependencies = [
+    "requests",
+    "psutil",
+    "tabulate"
+]
+
+
+[project.scripts]
+fc-devhost = "fc.devhost:main"
